+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- conversion of microtonal `WesternPitch`
+
+
 ## [0.15.0] - 2023-01-08
 
 ### Added